Split bitmap providers into one per file
