+++ lib/private/preview/bitmap.php
@@ -5,81 +5,33 @@
  * later.
  * See the COPYING-README file.
  */
+
 namespace OC\Preview;
 
-use Imagick;
+abstract class Bitmap extends Provider {
+	/**
+	 * {@inheritDoc}
+	 */
+	public function getThumbnail($path, $maxX, $maxY, $scalingup, $fileview) {
+		$tmpPath = $fileview->toTmpFile($path);
 
-	class Bitmap extends Provider {
+		//create imagick object from bitmap or vector file
+		try {
+			// Layer 0 contains either the bitmap or
+			// a flat representation of all vector layers
+			$bp = new \Imagick($tmpPath . '[0]');
 
-		public function getMimeType() {
-			return null;
+			$bp->setImageFormat('png');
+		} catch (\Exception $e) {
+			\OC_Log::write('core', $e->getmessage(), \OC_Log::ERROR);
+			return false;
 		}
 
-		public function getThumbnail($path, $maxX, $maxY, $scalingup, $fileview) {
-			$tmpPath = $fileview->toTmpFile($path);
+		unlink($tmpPath);
 
-			//create imagick object from bitmap or vector file
-			try {
-				// Layer 0 contains either the bitmap or
-				// a flat representation of all vector layers
-				$bp = new Imagick($tmpPath . '[0]');
-
-				$bp->setImageFormat('png');
-			} catch (\Exception $e) {
-				\OC_Log::write('core', $e->getmessage(), \OC_Log::ERROR);
-				return false;
-			}
-
-			unlink($tmpPath);
-
-			//new bitmap image object
-			$image = new \OC_Image($bp);
-			//check if image object is valid
-			return $image->valid() ? $image : false;
-		}
+		//new bitmap image object
+		$image = new \OC_Image($bp);
+		//check if image object is valid
+		return $image->valid() ? $image : false;
 	}
-
-		//.pdf
-		class PDF extends Bitmap {
-
-			public function getMimeType() {
-				return '/application\/pdf/';
-			}
-
-		}
-
-		//.tiff
-		class TIFF extends Bitmap {
-
-			public function getMimeType() {
-				return '/image\/tiff/';
-			}
-
-		}
-
-		//.ai
-		class Illustrator extends Bitmap {
-
-			public function getMimeType() {
-				return '/application\/illustrator/';
-			}
-
-		}
-
-		//.eps
-		class Postscript extends Bitmap {
-
-			public function getMimeType() {
-				return '/application\/postscript/';
-			}
-
-		}
-
-		//.psd
-		class Photoshop extends Bitmap {
-
-			public function getMimeType() {
-				return '/application\/x-photoshop/';
-			}
-
-		}
+}
